My GeoData Project
This file contains:

- destinations.data file: It has our data. In this case we have some
cities and their countries.

- geoload.py file : This file creates our database. Afterwards connects
with the GeoCode API of Google to retrieve information about a
city,country(destination).

- geodump.py file : This code reads our data from the database and
creates the where.js file. Afterwards we load this .js file to the
where.html file in order to visualize our data in a web browser.

-where.html file : This file loads the map in a web browser in order to
visualize our data.